+++ src/world/world.ts
@@ -144,12 +144,13 @@ const collectSunlightInvalidationWorldTileXRangeForLocalTile = (
   let maxWorldTileX = worldTileX + localLightingRange;
 
   if (localLightingRange === 0) {
+    minWorldTileX -= 1;
+    maxWorldTileX += 1;
+
     if (localX === 0) {
-      minWorldTileX -= 2;
-      maxWorldTileX += 1;
-    } else if (localX === CHUNK_SIZE - 1) {
       minWorldTileX -= 1;
-      maxWorldTileX += 2;
+    } else if (localX === CHUNK_SIZE - 1) {
+      maxWorldTileX += 1;
     }
   }